requirements.txt
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
dataclasses==0.8
ecdsa==0.14.1
fastapi==0.63.0
h11==0.12.0
passlib==1.7.4
pkg-resources==0.0.0
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
PyJWT==2.0.0
python-jose==3.2.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.22
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.3
